% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{StatToolkit-package}
\alias{StatToolkit-package}
\title{Utilities for Teaching Statistics}
\description{
Utilities for Teaching Statistics
}
\author{
\email{jason@bryer.org}

\email{bruce.dudek@albany.edu}
}
\keyword{package}
\keyword{statistics}
\keyword{teaching}

