% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic Data}
\format{a data frame with 1309 ovservations of 14 variables.}
\source{
Department of Biostatistics, Vanderbilt University
        http://biostat.mc.vanderbilt.edu/wiki/pub/Main/DataSets/titanic.html
}
\description{
This data set is used in the \code{GapAnalysis} demo and is used to demonstrate
how the \code{likert} package handles a gap analysis.
}
\keyword{datasets}

